+++ CHANGELOG.md
@@ -16,7 +16,9 @@ This release significantly improves the runtime performance of Metro-generated g
 - **Enhancement**: For multibindings that would use `Provider` elements/values, mark their sources as provider-accessed in refcounting.
 - **Enhancement**: Don't generate provider fields for graph self instances unless necessary.
 - **Enhancement**: Improve accuracy of diagnostic location when reporting graph validation issues from binding callable declarations.
+- **Enhancement**: Validate bindings from interop with Dagger `@BindsOptionalOf` lazily.
 - **Fix**: Fix `newInstance()` args not stripping `Lazy` in top-level function inject classes.
+- **Fix**: Allow `Any` to be a `binding<...>()` type if no explicit supertypes are declared.
 
 0.8.2
 -----